almost done with the extras part

+++ src/main/java/com/mindhub/homebanking/controller/AccountController.java
@@ -1,5 +1,5 @@
 package com.mindhub.homebanking.controller;
-
+import com.lowagie.text.*;
 import com.mindhub.homebanking.dtos.AccountDTO;
 import com.mindhub.homebanking.dtos.ClientDTO;
 import com.mindhub.homebanking.models.Account;
@@ -18,8 +18,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
+import java.awt.*;
+import java.io.FileOutputStream;
+import java.io.IOException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Random;
 import java.util.Set;
@@ -45,9 +49,51 @@ public class AccountController {
     @Autowired
     TransactionService transactionService;
 
-    @RequestMapping("/accounts/{id}")
-    public AccountDTO getAccount(@PathVariable Long id) {
-       return accountService.getAccountDTO(accountService.getAccount(id));
+    @GetMapping("/accounts/{id}")
+    public AccountDTO getAccount(@PathVariable Long id, @RequestParam("from_date") String from_date,
+                                 @RequestParam("thru_date") String thru_date) {
+        Account account = accountService.getAccount(id);
+        System.out.println(from_date);
+        System.out.println(thru_date);
+       if(!from_date.isEmpty() && !thru_date.isEmpty()) {
+           DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.SSSSSS");
+           LocalDateTime start_date = LocalDateTime.parse(from_date, formatter);
+           LocalDateTime end_date = LocalDateTime.parse(thru_date, formatter);
+           List<Transaction> transactions = transactionService.getFilteredTransactions(start_date, end_date, id);
+           account.setTransactions(transactions.stream().collect(Collectors.toSet()));
+           accountService.save(account);
+       }
+        Document document = new Document(PageSize.A4);
+        document.open();
+        document.add(new Paragraph("All transactions"));
+        Table table = new Table(3);
+        table.setBorderWidth(1);
+        table.setBorderColor(new Color(0, 0, 255));
+        table.setPadding(5);
+        table.setSpacing(5);
+        Cell cell = new Cell("header");
+        cell.setHeader(true);
+        cell.setColspan(3);
+        table.addCell(cell);
+        table.endHeaders();
+        cell = new Cell("example cell with colspan 1 and rowspan 2");
+        cell.setRowspan(2);
+        cell.setBorderColor(new Color(255, 0, 0));
+        table.addCell(cell);
+        table.addCell("1.1");
+        table.addCell("2.1");
+        table.addCell("1.2");
+        table.addCell("2.2");
+        table.addCell("cell test1");
+        cell = new Cell("big cell");
+        cell.setRowspan(2);
+        cell.setColspan(2);
+        table.addCell(cell);
+        table.addCell("cell test2");
+        document.add(table);
+        document.close();
+
+        return accountService.getAccountDTO(account);
     }
     @RequestMapping("/accounts")
     public List<AccountDTO> getAccounts() {
@@ -81,12 +127,17 @@ public class AccountController {
     }
 
     @PatchMapping("/clients/current/accounts/{id}")
-    public void disableAccount(@PathVariable Long id) {
+    public ResponseEntity<Object> disableAccount(@PathVariable Long id) {
         Account account = accountService.getAccount(id);
-        account.setEnabled(false);
-        Set<Transaction> transactions = account.getTransactions();
-        transactions.forEach(transaction -> transactionService.deleteTransaction(transaction.getId()));
-        accountService.save(account);
+        if(account.getBalance() > 0) {
+            return new ResponseEntity<>("Cannot delete accounts that have money", HttpStatus.FORBIDDEN);
+        } else {
+            account.setEnabled(false);
+            Set<Transaction> transactions = account.getTransactions();
+            transactions.forEach(transaction -> transactionService.deleteTransaction(transaction.getId()));
+            accountService.save(account);
+        }
+        return new ResponseEntity<>("Erased", HttpStatus.ACCEPTED);
     }
